+++ src/neural_network/optimizer.rs
@@ -5,3 +5,8 @@ pub mod sgd;
 pub use adam::*;
 pub use rmsprop::*;
 pub use sgd::*;
+
+pub struct OptimizerCache {
+    pub adam_states: Option<AdamStates>,
+    pub rmsprop_cache: Option<RMSpropCache>,
+}